Refactor TableStorageService
